Update correct type for the time to ensure the precision is not lost

+++ code/pir_event.py
@@ -9,7 +9,7 @@ class PirEventType(Enum):
 
 class PIREvent:
     """Class for defining the messages for the PIR Queue in PIR_Sensor class"""
-    time: int
+    time: float
     event_type: PirEventType
 
     def __init__(self, time: float, event_type: PirEventType):
